Provide logger instance to FriendService. Rename source files to maintain consistency.
